#15 리뷰반영 1 - user 패키지 분리(authentication 패키지 생성), global 패키지 생성
