+++ src/main/java/com/flab/doorrush/global/config/SecurityConfig.java
@@ -1,4 +1,4 @@
-package com.flab.doorrush.config;
+package com.flab.doorrush.global.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -6,7 +6,7 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 
 @Configuration
-public class SecurityConfigurer {
+public class SecurityConfig {
 
   @Bean
   public PasswordEncoder passwordEncoder() {
